There is no llm json command
I was 'vibe-coding' this tutorial and didn't fact-check it.
Thanks, Carlton. Rewrite the llm json example with --schema-multi

Is this a signal that I should check more?
No, I don't think so. I was making as many mistakes myself.
So my lesson is that the LLM is no worse than me.
Let's vibe code more and check less! :-)