Update README.md
Extended to 10th Aug, which is normally when assignments are traditionally due, just to give them a little more time to test with our endpoint testing mechanism which will only be available by tomorrow. That way if we have any bugs in our deployment we will have time to fix it as well.